Add rate model, viewmodel and view. Add migration to price in rate table.

+++ PublicServicesDomain/Models/Rate.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -11,7 +12,7 @@ namespace PublicServicesDomain.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string MeasureTitle { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public User User { get; set; }
     }
 }
